Refactor executive summary helpers
